Rename DebugDraw.h/cc to DebugDrawList to match the class name
