+++ core/CMakeLists.txt
@@ -57,7 +57,7 @@ set(MONDRADIKO_CORE_SRC
   gpu/GpuPipeline.cc
   gpu/GpuShader.cc
   physics/Physics.cc
-  renderer/DebugDraw.cc
+  renderer/DebugDrawList.cc
   renderer/MeshPass.cc
   renderer/OverlayPass.cc
   renderer/Renderer.cc
