+++ core/renderer/DebugDrawList.cc
@@ -1,7 +1,7 @@
 // Copyright (c) 2020-2021 the Mondradiko contributors.
 // SPDX-License-Identifier: LGPL-3.0-or-later
 
-#include "core/renderer/DebugDraw.h"
+#include "core/renderer/DebugDrawList.h"
 
 #include <utility>
